be more flexible about data dictionary version
users may not always enter the data dictionary version argument exactly
as specified. be flexible and try to figure out which major and minor
versions they asked for. if they did not specify a major and minor
versions that are currently supported, throw the usual error for not
making an allowed choice.