Versions for 2026 (#46)
* be more flexible about data dictionary version

users may not always enter the data dictionary version argument exactly
as specified. be flexible and try to figure out which major and minor
versions they asked for. if they did not specify a major and minor
versions that are currently supported, throw the usual error for not
making an allowed choice.

* update effective and enforcement dates for each version

* bumping htp validator version

* allowed for the hpt validator to be pulled from npm vs github

---------

Co-authored-by: Mint Thompson <mathompson@mitre.org>